projects/Chatbot-Based Employee Management System/requirements.txt
flask
pymongo
psycopg2-binary
transformers